+++ CHANGELOG.rst
@@ -4,6 +4,13 @@ Changelog
 ---------
 
 
+0.3.8 (2017-12-?)
+^^^^^^^^^^^^^^^^^
+
+* add anti-captcha.com support (all credits go to kirov #353)
+* add futbinPrice, futheadPrice
+* add missing params to club method (#351)
+
 0.3.7 (2017-12-01)
 ^^^^^^^^^^^^^^^^^^